Updated Window Size With regualted physics

+++ PA_9/PA_9/source.cpp
@@ -1,6 +1,6 @@
 #include "source.hpp"
 
-double getVectorManitude(sf::Vector2f vector)
+float getVectorManitude(sf::Vector2f vector)
 {
 	return sqrt(vector.x * vector.x + vector.y * vector.y);
 }
